// packages/b_declarations/src/properties/font-variant-numeric/types.ts
// b_path:: packages/b_declarations/src/properties/font-variant-numeric/types.ts

export type FontVariantNumericIR = {
  kind: "keyword";
  value:
    | "normal"
    | "lining-nums"
    | "oldstyle-nums"
    | "proportional-nums"
    | "tabular-nums"
    | "diagonal-fractions"
    | "stacked-fractions"
    | "ordinal"
    | "slashed-zero"
    | "inherit"
    | "initial"
    | "revert"
    | "revert-layer"
    | "unset";
};